pyproject.toml
[project]
name = "sopra"
version = "0.1.0"
description = "Enterprise CRUD demo with Streamlit, pandas, SQLAlchemy, and Microsoft SQL Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=3.0.2",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "streamlit>=1.56.0",
]
